Extracted OspreySharp calibration into a Calibrator collaborator
* Moved the RT/mass calibration cluster (12 methods, CalibrationPassResult, 5 constants) out of PerFileScoringTask into new Tasks/Calibrator.cs; PerFileScoringTask 2828 -> 1558 LOC
* Promoted ExtractTopNFragmentXics/FindNearestMs1/CountTop6Matches to internal static so the standalone Calibrator can call them
* Byte-identical relocation: build net8.0+net472, 372 tests, 0 warnings; Stellar+Astral cross-impl parity PASS at 1e-9

See ai/todos/active/TODO-20260607_ospreysharp_extract_calibrator.md
